fix: Complete ZTCF/ZVCF stubs for all engines + mypy fixes
All remaining mypy errors resolved:
- ✅ Drake: Added ZTCF/ZVCF stubs (2/5)
- ✅ Pinocchio: Added ZTCF/ZVCF stubs (3/5)
- ✅ OpenSim: Added ZTCF/ZVCF stubs (4/5)
- ✅ MyoSuite: Added ZTCF/ZVCF stubs (5/5)
- ✅ test_pendulum_lagrangian.py: Cast returns to float (fix no-any-return)

All 5 engines now have stub implementations that:
- Raise NotImplementedError with helpful message
- Reference pendulum_physics_engine.py for implementation example
- Satisfy PhysicsEngine protocol abstract method requirements

CI should now pass all checks (black, ruff, mypy)!

+++ tests/analytical/test_pendulum_lagrangian.py
@@ -73,7 +73,7 @@ class TestPendulumAnalyticalDynamics:
         """
         I_inertia = self.m1 * self.l1**2  # Inertia [kg·m²]
         gravity_torque = self.m1 * self.g * self.l1 * np.sin(theta)
-        return I_inertia * theta_ddot + gravity_torque
+        return float(I_inertia * theta_ddot + gravity_torque)
 
     def test_inverse_dynamics_vertical_down(self) -> None:
         """Test ID at vertical down (θ=0): pure inertial torque.
@@ -299,7 +299,7 @@ class TestPendulumEnergyConservation:
         I_inertia = self.m1 * self.l1**2
         KE = 0.5 * I_inertia * theta_dot**2
         PE = self.m1 * self.g * self.l1 * (1 - np.cos(theta))
-        return KE + PE
+        return float(KE + PE)
 
     def test_energy_conservation_free_swing(self) -> None:
         """Test energy is conserved during passive swing (no torque).
